# server/api/openapi.bundled.yaml
openapi: 3.1.0
info:
  title: StuHelper API
  version: 1.0.0
  description: StuHelper 学习助手平台 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:8080
    description: Development
  - url: https://api.stuhelper.com
    description: Production
...
paths:
...
  /api/v1/course/review/courses/{courseID}/reviews:
    get:
      tags:
        - Review
      operationId: getCourseReviews
      summary: 获取课程测评列表（支持可选认证）
      security:
        - {}
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseIDPath'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: sort
          in: query
          schema:
            type: string
            enum:
              - time
              - likes
              - rating
            default: time
        - name: termID
          in: query
          schema:
            type: string
          description: 按学期筛选
        - name: teacherID
          in: query
          schema:
            type: integer
            format: int64
          description: 按教师筛选
      responses:
        '200':
          description: 测评分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: object
                        required:
                          - list
                          - total
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/Review'
                          total:
                            type: integer
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/course/review/reviews:
    post:
      tags:
        - Review
      operationId: postReview
      summary: 发布测评
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostReviewRequest'
      responses:
        '201':
          description: 测评发布成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/course/review/reviews/{reviewID}:
    put:
      tags:
        - Review
      operationId: updateReview
      summary: 更新测评
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageData'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags:
        - Review
      operationId: deleteReview
      summary: 删除测评
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewIDPath'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageData'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/course/review/user/reviews:
    get:
      tags:
        - User
      operationId: getUserReviews
      summary: 获取用户评论列表
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 用户评论分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: object
                        required:
                          - list
                          - total
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/Review'
                          total:
                            type: integer
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/course/review/admin/reports:
    get:
      tags:
        - Admin
      operationId: listReports
      summary: 获取举报列表
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              - pending
              - resolved
              - rejected
              - all
            default: pending
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 举报分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: object
                        required:
                          - list
                          - total
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/ReviewReport'
                          total:
                            type: integer
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/course/review/admin/reports/{reportID}:
    put:
      tags:
        - Admin
      operationId: processReport
      summary: 处理举报
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessReportRequest'
      responses:
        '200':
          description: 处理成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageData'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/course/review/admin/reviews:
    get:
      tags:
        - Admin
      operationId: listAllReviews
      summary: 获取所有评论（管理员）
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              - published
              - hidden
              - deleted
              - all
            default: all
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 评论分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: object
                        required:
                          - list
                          - total
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/Review'
                          total:
                            type: integer
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/course/review/admin/reviews/{reviewID}:
    put:
      tags:
        - Admin
      operationId: adminUpdateReview
      summary: 管理员更新评论状态
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateReviewRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageData'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/course/review/admin/reviews/batch:
    post:
      tags:
        - Admin
      operationId: batchUpdateReviews
      summary: 批量更新评论状态
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchUpdateReviewsRequest'
      responses:
        '200':
          description: 批量更新完成
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/BatchUpdateResult'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/course/review/admin/stats:
    get:
      tags:
        - Admin
      operationId: getAdminStats
      summary: 获取管理统计
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: 管理统计数据
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/AdminStats'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: 统一错误响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseBody'
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          enum:
            - true
    ErrorResponseBody:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          $ref: '#/components/schemas/APIError'
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: A00400
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    MessageData:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    Review:
      type: object
      required:
        - id
        - courseID
        - title
        - termID
        - content
        - ratings
        - likeCount
        - dislikeCount
        - replyCount
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        courseID:
          type: integer
          format: int64
        courseName:
          type: string
        teacherID:
          type:
            - integer
            - 'null'
          format: int64
        teacherName:
          type: string
        termID:
          type: string
        termName:
          type: string
        title:
          type: string
        content:
          type: string
        grade:
          type: string
        ratings:
          $ref: '#/components/schemas/ReviewRatings'
        likeCount:
          type: integer
        dislikeCount:
          type: integer
        replyCount:
          type: integer
        status:
          type: string
          enum:
            - published
            - hidden
            - deleted
        moderationReason:
          type:
            - string
            - 'null'
          description: 屏蔽原因（仅 hidden 状态时存在）
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReviewRatings:
      type: object
      additionalProperties:
        type: integer
        minimum: 1
        maximum: 5
      description: Map of dimension key to rating value (1-5)
    PostReviewRequest:
      type: object
      required:
        - courseID
        - title
        - termID
        - content
        - ratings
      properties:
        courseID:
          type: integer
          format: int64
          minimum: 1
        teacherID:
          type: integer
          format: int64
          minimum: 1
        termID:
          type: string
          maxLength: 20
          pattern: ^\d{4}-[12]$
        title:
          type: string
          maxLength: 200
        content:
          type: string
          minLength: 10
          maxLength: 5000
        grade:
          type: string
        ratings:
          $ref: '#/components/schemas/ReviewRatings'
    UpdateReviewRequest:
      type: object
      required:
        - content
        - ratings
      properties:
        title:
          type: string
          maxLength: 200
        content:
          type: string
          minLength: 10
          maxLength: 5000
        grade:
          type: string
        ratings:
          $ref: '#/components/schemas/ReviewRatings'
...
    ReviewReport:
      type: object
      required:
        - id
        - reviewID
        - reason
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        reviewID:
          type: string
          format: uuid
        review:
          $ref: '#/components/schemas/Review'
        reason:
          type: string
          enum:
            - spam
            - inappropriate
            - harassment
            - false_info
            - other
        description:
          type: string
        status:
          type: string
          enum:
            - pending
            - resolved
            - rejected
        resolvedBy:
          type:
            - string
            - 'null'
        resolvedAt:
          type:
            - string
            - 'null'
          format: date-time
        resolutionNote:
          type:
            - string
            - 'null'
        createdAt:
          type: string
          format: date-time
    ProcessReportRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum:
            - reject
            - hide
            - delete
        note:
          type: string
          maxLength: 500
    AdminUpdateReviewRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum:
            - hide
            - restore
            - delete
        reason:
          type: string
          maxLength: 500
    BatchUpdateReviewsRequest:
      type: object
      required:
        - ids
        - action
      properties:
        ids:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
          maxItems: 100
        action:
          type: string
          enum:
            - hide
            - restore
            - delete
    BatchUpdateResult:
      type: object
      required:
        - message
        - affected
      properties:
        message:
          type: string
        affected:
          type: integer
          format: int64
    AdminStats:
      type: object
      required:
        - totalReviews
        - publishedReviews
        - hiddenReviews
        - deletedReviews
        - pendingReports
        - totalReports
        - todayReviews
        - weekReviews
      properties:
        totalReviews:
          type: integer
        publishedReviews:
          type: integer
        hiddenReviews:
          type: integer
        deletedReviews:
          type: integer
        pendingReports:
          type: integer
        totalReports:
          type: integer
        todayReviews:
          type: integer
        weekReviews:
          type: integer
...
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: 每页数量
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    CourseIDPath:
      name: courseID
      in: path
      required: true
      description: 课程 ID
      schema:
        type: integer
        format: int64
    ReviewIDPath:
      name: reviewID
      in: path
      required: true
      description: 测评 ID
      schema:
        type: string
        format: uuid
...
    ReportIDPath:
      name: reportID
      in: path
      required: true
      description: 举报 ID
      schema:
        type: string
        format: uuid
...